Update dependencies and improve code readability
The commit updates various dependencies in the package-lock.json file, including those of chromium-bidi, devtools-protocol, puppeteer, and zod. This ensures the software is up-to-date. The code in index.js, browser.js, and scrape.js have also been updated for improved code readability and consistency, making it easier to understand and maintain the codebase. Some variables were renamed for clearer naming conventions.